Added confirmation box to shutdown script
Added confirmation box to shutdown script to make sure system doesn't shutdown immediately if shortcut keys collide with some other apps

+++ shutdown.sh
@@ -1,6 +1,11 @@
 umount /media/pavan/dreams3
 if [ $? -eq 0 ]; then
-	poweroff
+	sure=$(zenity --entry --text "sure? press y to continue")
+	if [ $sure = y ]; then 
+		poweroff
+	else 
+		echo "ok"
+	fi
 else
 	zenity --error --text="drive is busy!"
 fi
